Implement persistent BF16 KV cache for Talker and Code Predictor
Store KV cache entries as bf16 (uint16_t) instead of f32, halving
memory usage. Key changes:

- KV cache fields changed from float* to uint16_t* in qwen_tts_ctx_t
- f32→bf16 conversion (NEON-vectorized) when writing K/V to cache
- New qwen_causal_attention_bf16kv() kernel that reads bf16 K/V with
  inline NEON conversion during dot products and V accumulation
- Prefill uses f32 Q/K/V directly for attention (avoids bf16 roundtrip),
  then stores bf16 to cache for subsequent decode steps
- KV cache grow/alloc updated for uint16_t element size

This improves cache line utilization for the bandwidth-bound single-token
decode path, where attention reads the full KV history each step.

+++ qwen_tts.h
@@ -367,15 +367,15 @@ typedef struct qwen_tts_ctx {
     /* Speaker encoder (ECAPA-TDNN, Base model only) */
     qwen_speaker_encoder_t speaker_enc;
     
-    /* KV cache (Talker) */
-    float *kv_cache_k;
-    float *kv_cache_v;
+    /* KV cache (Talker) — stored as bf16 to halve memory and improve cache utilization */
+    uint16_t *kv_cache_k;
+    uint16_t *kv_cache_v;
     int kv_max;
     int kv_len;
-    
-    /* KV cache (Code Predictor) */
-    float *cp_kv_k;
-    float *cp_kv_v;
+
+    /* KV cache (Code Predictor) — stored as bf16 */
+    uint16_t *cp_kv_k;
+    uint16_t *cp_kv_v;
     int cp_kv_max;
     int cp_kv_len;
